+++ apps/namadillo/src/App/Layout/NamadaAccount.tsx
@@ -1,0 +1,49 @@
+import { CopyToClipboardControl, Tooltip } from "@namada/components";
+import { routes } from "App/routes";
+import { defaultAccountAtom, disconnectAccountAtom } from "atoms/accounts";
+import { useAtomValue } from "jotai";
+import { useLocation, useNavigate } from "react-router-dom";
+import { DisconnectAccountIcon } from "./DisconnectAccountIcon";
+import { SwitchAccountIcon } from "./SwitchAccountIcon";
+import NamadaWalletIcon from "./assets/namada-wallet-icon.svg";
+
+export const NamadaAccount = (): JSX.Element => {
+  const { data: account, isFetching } = useAtomValue(defaultAccountAtom);
+  const { mutateAsync: disconnect } = useAtomValue(disconnectAccountAtom);
+  const location = useLocation();
+  const navigate = useNavigate();
+
+  if (!account || isFetching) {
+    return <></>;
+  }
+
+  const buttonClassName = "p-1 opacity-80 transition-opacity hover:opacity-100";
+
+  return (
+    <div className="flex items-center rounded-sm text-sm text-white bg-rblack px-1">
+      <span className="flex items-center gap-2 relative group/tooltip">
+        <CopyToClipboardControl
+          className={buttonClassName}
+          value={account.address || ""}
+        >
+          <img src={NamadaWalletIcon} alt="Namada Wallet" className="w-6 h-6" />
+          <span className="mr-1">{account.alias}</span>
+        </CopyToClipboardControl>
+        <Tooltip position="left">{account.address}</Tooltip>
+      </span>
+      <button
+        className={buttonClassName}
+        onClick={() => {
+          navigate(routes.switchAccount, {
+            state: { backgroundLocation: location },
+          });
+        }}
+      >
+        <SwitchAccountIcon />
+      </button>
+      <button className={buttonClassName} onClick={() => disconnect()}>
+        <DisconnectAccountIcon />
+      </button>
+    </div>
+  );
+};
