+++ apps/namadillo/src/App/Layout/TopNavigation.tsx
@@ -10,7 +10,8 @@ import { useAtomValue } from "jotai";
 import { AiOutlineMessage } from "react-icons/ai";
 import { IoSettingsOutline } from "react-icons/io5";
 import { useLocation, useNavigate } from "react-router-dom";
-import { ActiveAccount } from "./ActiveAccount";
+import { KeplrAccount } from "./KeplrAccount";
+import { NamadaAccount } from "./NamadaAccount";
 import { SyncIndicator } from "./SyncIndicator";
 
 export const TopNavigation = (): JSX.Element => {
@@ -82,7 +83,10 @@ export const TopNavigation = (): JSX.Element => {
 
       <SyncIndicator />
 
-      <ActiveAccount />
+      <div className="h-[50px] flex gap-1">
+        <NamadaAccount />
+        <KeplrAccount />
+      </div>
     </div>
   );
 };
